+++ pyfreebilling/endpoint/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class EndpointConfig(AppConfig):
+    name = 'pyfreebilling.endpoint'
+    verbose_name = "endpoint"